+++ app/src/main/java/com/rob729/movies/models/network/MovieListResponse.kt
@@ -1,7 +1,7 @@
 package com.rob729.movies.models.network
 
 import androidx.annotation.Keep
-import com.rob729.movies.Constants
+import com.rob729.movies.utils.Constants
 import com.rob729.movies.models.database.MovieData
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
@@ -15,7 +15,7 @@ data class MovieListResponse(
 @Keep
 @JsonClass(generateAdapter = true)
 data class MovieItem (
-    @Json(name = "id") val id: Int? = null,
+    @Json(name = "id") val id: Long? = null,
     @Json(name = "title") val title: String? = null,
     @Json(name = "poster_path") val posterPath: String? = null,
     @Json(name = "overview") val overview: String? = null,
